+++ src/extension.ts
@@ -1,27 +1,29 @@
-import { relative, sep } from 'path';
 import { commands, ExtensionContext, window, workspace } from 'vscode';
 
-import { FileService } from './files-service/file.service';
+import { FileService, QuoteType } from './files-service/file.service';
 
 
 export function activate(context: ExtensionContext) {
 	let disposable = commands.registerCommand('extension.indexerator', (fullPath) => {
-		// get file name from settings- index as default file name.
+		// get file name from user settings- index as default file name.
 		const fileName = workspace.getConfiguration().get('indexerator.exportFileName') as string;
+		// get quote style from user settings
+		const quoteStyle = workspace.getConfiguration().get('indexerator.quotes') as QuoteType;
 		const fileService = new FileService(fileName);
 
-		fileService.on('error', () => {
-				window.showErrorMessage(`Unable to create ${fileService.indexFileName}`)
-			})
-			.on('create', () => {
-				const folderName = relative(process.cwd(), fullPath.fsPath).split(sep);
-				window.setStatusBarMessage(`indexerator >index in ${folderName[folderName.length - 1]}`);
+		fileService.setQuoteStyle(quoteStyle);
 
-				// clear status bar after 5 sec
-				setTimeout(() => window.setStatusBarMessage(''), 5 * 1000);
-			});
+		fileService.on('error', (reason: string) => {
+			window.showErrorMessage(`indexerator: ${reason}`);
+		})
+		.on('create', (reason) => {
+			window.setStatusBarMessage(`indexerator >${reason}`);
 
-			fileService.generateExportFile(fullPath.fsPath);
+			// clear status bar after 5 sec
+			setTimeout(() => window.setStatusBarMessage(''), 5 * 1000);
+		});
+
+		fileService.generateExportFile(fullPath.fsPath);
 	});
 
 	context.subscriptions.push(disposable);